+++ src/components/TictactoeComponents/BoardButton.tsx
@@ -1,11 +1,12 @@
 interface BoardButtonType {
 	value: string;
+	isDone: boolean;
 	onClick: (e: React.MouseEvent<HTMLButtonElement>) => void;
 }
 
-export const BoardButton = ({ value, onClick }: BoardButtonType) => {
+export const BoardButton = ({ value, isDone, onClick }: BoardButtonType) => {
 	return (
-		<button className='boardButton' onClick={onClick}>
+		<button className='boardButton' disabled={isDone} onClick={onClick}>
 			{value}
 		</button>
 	);
