+++ src/components/TictactoeComponents/GridGameBoard.tsx
@@ -1,3 +1,4 @@
+import useCheckWinner from '../../hooks/useCheckWinner';
 import { BoardButton } from './BoardButton';
 
 interface GridGameBoardType {
@@ -6,10 +7,13 @@ interface GridGameBoardType {
 }
 
 export const GridGameBoard = ({ board, onClick }: GridGameBoardType) => {
+	// 현재 승자를 체크
+	const winner = useCheckWinner(board);
+
 	return (
 		<div className='GridGameBoard'>
 			{board.map((value, index) => (
-				<BoardButton key={index} value={value} onClick={() => onClick(index)} />
+				<BoardButton key={index} value={value} isDone={winner === 'O' || winner === 'X' || winner === 'draw' ? true : false} onClick={() => onClick(index)} />
 			))}
 		</div>
 	);
