+++ src/styles/tictactoe.scss
@@ -4,33 +4,43 @@
 	color: #eee;
 	padding: 60px 0;
 	margin: 0 auto;
+
 	> h2 {
 		font-size: 22px;
 		font-weight: lighter;
 		color: #888;
 		margin-bottom: 20px;
 	}
+
 	> p {
 		font-size: 20px;
 		color: #fefefe;
 		margin-bottom: 20px;
+
 		> span {
 			&.player1 {
 				color: red;
 			}
+
 			&.player2 {
 				color: blue;
 			}
 		}
 	}
+
 	.gridGameBoard {
 		margin-bottom: 20px;
 	}
+
 	button.reset {
 		background-color: #fff022;
 		border: 1px solid #666;
 		border-radius: 3px;
 		padding: 5px 10px;
+	}
+
+	.status {
+		margin-top: 20px;
 	}
 }
 
@@ -41,20 +51,10 @@
 	grid-template-rows: repeat(3, 100px);
 	gap: 5px;
 	margin-bottom: 20px;
-}
 
-.boardButton {
-	width: 100px;
-	height: 100px;
-	font-size: 3rem;
-}
-
-.status {
-	margin-top: 20px;
-}
-
-button {
-	font-size: 1rem;
-	padding: 10px 20px;
-	margin-top: 10px;
+	.boardButton {
+		width: 100px;
+		height: 100px;
+		font-size: 3rem;
+	}
 }
